+++ lbrynet/daemon/testMedia.py
@@ -1,0 +1,3 @@
+from MediaDescriptor import get_video_length
+
+get_video_length("/home/kamilp/Downloads/Kamil P. TR-909 ROLAND GAIA MINIBRUTE_1.mp4")